update :: 확장자 validation이 추가되면서 중복되는 코드를 추상 클래스에서 관리하도록 변경
